requirements.txt
# Core ML (STRICT versions - do not change)
tensorflow==2.15.0
numpy==1.26.4

# Backend
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9

# Image processing
opencv-python==4.8.1.78
pillow==10.2.0
matplotlib==3.8.2

# ML utilities
scikit-learn==1.3.2

# RAG / NLP
faiss-cpu==1.7.4
sentence-transformers==2.6.1
transformers==4.38.2

# Utils
requests==2.31.0
python-dotenv==1.0.1

# Optional (keep if using search APIs)
google-search-results==2.4.2
serpapi==0.1.5

ultralytics
openai
requests
seaborn